Split out pen tool options into another file
